+++ include/dynamics/ODE.hpp
@@ -9,10 +9,10 @@ public:
     virtual ~ODE() = default;
 
     virtual void derivatives(
-        const std::vector<Eigen::Vector3f>& positions,
-        const std::vector<Eigen::Vector3f>& velocities,
-        std::vector<Eigen::Vector3f>& dpos_dt,
-        std::vector<Eigen::Vector3f>& dvel_dt
+        const std::vector<Eigen::Vector3d>& positions,
+        const std::vector<Eigen::Vector3d>& velocities,
+        std::vector<Eigen::Vector3d>& dpos_dt,
+        std::vector<Eigen::Vector3d>& dvel_dt
     ) = 0;
 };
