+++ src/dynamics/CR3BP.cpp
@@ -3,15 +3,15 @@
 
 namespace Dynamics {
 
-CR3BPModel::CR3BPModel(float mu_)
+CR3BPModel::CR3BPModel(double mu_)
     : mu(mu_)
 {}
 
 void CR3BPModel::derivatives(
-    const std::vector<Eigen::Vector3f>& positions,
-    const std::vector<Eigen::Vector3f>& velocities,
-    std::vector<Eigen::Vector3f>& dpos_dt,
-    std::vector<Eigen::Vector3f>& dvel_dt
+    const std::vector<Eigen::Vector3d>& positions,
+    const std::vector<Eigen::Vector3d>& velocities,
+    std::vector<Eigen::Vector3d>& dpos_dt,
+    std::vector<Eigen::Vector3d>& dvel_dt
 ) {
 
     int n = positions.size();
@@ -20,8 +20,8 @@ void CR3BPModel::derivatives(
     dvel_dt.resize(n);
 
     for (int i = 0; i < n; i++) {
-        dpos_dt[i] = Eigen::Vector3f::Zero();
-        dvel_dt[i] = Eigen::Vector3f::Zero();
+        dpos_dt[i] = Eigen::Vector3d::Zero();
+        dvel_dt[i] = Eigen::Vector3d::Zero();
     }
 
     if (n < 1) return;
@@ -29,34 +29,34 @@ void CR3BPModel::derivatives(
     const auto& r = positions[0];
     const auto& v = velocities[0];
 
-    float x = r.x();
-    float y = r.y();
-    float z = r.z();
+    double x = r.x();
+    double y = r.y();
+    double z = r.z();
 
-    float vx = v.x();
-    float vy = v.y();
-    float vz = v.z();
+    double vx = v.x();
+    double vy = v.y();
+    double vz = v.z();
 
-    float r1 = std::sqrt((x + mu)*(x + mu) + y*y + z*z);
-    float r2 = std::sqrt((x - (1 - mu))*(x - (1 - mu)) + y*y + z*z);
+    double r1 = std::sqrt((x + mu)*(x + mu) + y*y + z*z);
+    double r2 = std::sqrt((x - (1 - mu))*(x - (1 - mu)) + y*y + z*z);
 
     dpos_dt[0] = v;
 
-    float ax =
-        2.0f * vy + x
+    double ax =
+        2.0 * vy + x
         - (1 - mu)*(x + mu)/std::pow(r1, 3)
         - mu*(x - (1 - mu))/std::pow(r2, 3);
 
-    float ay =
-        -2.0f * vx + y
+    double ay =
+        -2.0 * vx + y
         - (1 - mu)*y/std::pow(r1, 3)
         - mu*y/std::pow(r2, 3);
 
-    float az =
+    double az =
         - (1 - mu)*z/std::pow(r1, 3)
         - mu*z/std::pow(r2, 3);
 
-    dvel_dt[0] = Eigen::Vector3f(ax, ay, az);
+    dvel_dt[0] = Eigen::Vector3d(ax, ay, az);
 
 }
